backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
firecrawl-py
pydantic
python-dotenv
httpx
